# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: AI Homework Grader API
  version: "0.2.0"
  description: |
    API for AI-powered homework grading. Upload exam photos, receive annotated
    grading results, and browse error-question collections. MVP covers English
    and Math.

    ## Authentication (v0.2.0)
    Parent identity is established via SMS verification-code login
    (`POST /api/auth/send-code` + `POST /api/auth/login`), which issues a JWT.
    All business endpoints require `Authorization: Bearer <token>`; the JWT
    `sub` claim is the parent id (userId). Phone appears ONLY in the auth
    endpoints — never in business URLs, query params, or logs.

    Image endpoints use signed URLs (HMAC token + expiry in the query string)
    because `<img>`/`<image>` tags cannot send an Authorization header.

    The WeChat mini-program uses the SAME SMS login flow as the Web client
    (`POST /api/auth/send-code` + `POST /api/auth/login`); there is no
    mini-program-specific login endpoint. The issued JWT is cached client-side
    and reused until it expires.
servers:
  - url: http://localhost:8000
    description: Local development (FastAPI backend)

# Global security: all business endpoints require a Bearer JWT unless overridden.
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/send-code:
    post:
      operationId: sendVerificationCode
      tags: [Auth]
      security: []
      summary: Send an SMS verification code
      description: |
        Sends a 6-digit code via Aliyun SMS. Rate-limited: 60s between sends per
        phone. The code is valid for 5 minutes. The code itself is never
        returned in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: '^\d{11}$'
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                required: [retry_after]
                properties:
                  retry_after:
                    type: integer
                    example: 60
                    description: Seconds before a new code may be requested
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      security: []
      summary: Verify code and issue a JWT
      description: |
        Verifies the SMS code and issues a JWT whose `sub` claim is the parent
        id (userId). First login auto-creates the Parent (phone UNIQUE). This is
        the primary login path for both Web and mini-program.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                  pattern: '^\d{11}$'
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ═══════════════════════════════════════════════════════
  # Children
  # ═══════════════════════════════════════════════════════
  /api/children:
    get:
      operationId: listChildren
      tags: [Children]
      summary: List children for the authenticated parent
      responses:
        "200":
          description: List of children
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Child"

    post:
      operationId: createChild
      tags: [Children]
      summary: Add a child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                grade:
                  type: string
                  enum: [一年级, 二年级, 三年级, 四年级, 五年级, 六年级]
                  default: 五年级
                  description: Required on the data model; defaults to 五年级 when omitted.
                note:
                  type: string
                  maxLength: 200
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "409":
          description: Duplicate name for this parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/children/{child_id}:
    put:
      operationId: updateChild
      tags: [Children]
      summary: Update a child (name / grade / note)
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                grade:
                  type: string
                  enum: [一年级, 二年级, 三年级, 四年级, 五年级, 六年级]
                note:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: Child updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteChild
      tags: [Children]
      summary: Delete a child (history preserved)
      parameters:
        - name: child_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Child deleted, submissions preserved
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════════════
  # Submissions
  # ═══════════════════════════════════════════════════════
...
components:
  schemas:
...
    LoginResponse:
      type: object
      required: [token, token_type, expires_at, user_id]
      properties:
        token:
          type: string
          description: JWT (sub = parent id, exp = 30 days, HS256)
        token_type:
          type: string
          enum: [Bearer]
        expires_at:
          type: string
          format: date-time
        user_id:
          type: integer
          description: Parent id (userId)

    # ── Children ──────────────────────────────────────
    Child:
      type: object
      required: [id, name, grade, submission_count, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
          example: "小明"
        grade:
          type: string
          enum: [一年级, 二年级, 三年级, 四年级, 五年级, 六年级]
          description: Required; defaults to 五年级 on create.
        note:
          type: string
          nullable: true
          maxLength: 200
        avatar:
          type: string
          nullable: true
          description: Reserved field; not implemented in v2 (no upload/edit/display).
        submission_count:
          type: integer
          description: Total submissions for this child
        created_at:
          type: string
          format: date-time

    # ── Submissions ───────────────────────────────────
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

  # ── Security Schemes ──────────────────────────────
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by POST /api/auth/login. The `sub` claim is the parent id
        (userId). Applies to all business endpoints.
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Future scheme — retained for migration reference only. NOT implemented
        in v2.

  # ── Responses ───────────────────────────────────────
  responses:
...
    NotFound:
      description: Resource not found (also used for cross-owner access to avoid existence probing)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "Submission not found."
...